@@ -9,21 +9,21 @@ authors = ["ksdfg <ksdfg123@gmail.com>"]
 license = "GPL-3.0"
 
 [tool.poetry.dependencies]
-python = "3.9.5"
+python = "3.10.1"
 bs4 = "^0.0.1"
-dnspython = "^2.0.0"
-emoji = "^0.6.0"
-mongoengine = "^0.20.0"
-Pillow = "^8.0.1"
-python-decouple = "^3.3"
-python-telegram-bot = "^13.0"
+dnspython = "^2.1.0"
+emoji = "^1.6.1"
+mongoengine = "^0.23.1"
+Pillow = "^8.4.0"
+python-decouple = "^3.5"
+python-telegram-bot = "^13.9"
 spongemock = "^0.3.4"
 zalgo-text = "^0.6"
-hentai = "^3.2.5"
+hentai = "^3.2.9"
 telegraph = "^2.1.0"
 
 [tool.poetry.dev-dependencies]
-black = "^20.8b1"
+black = "^21.12b0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
